requirements.txt
beautifulsoup4==4.11.1
discord.py==2.0.1
lxml==4.9.1
numpy==1.23.4
requests==2.28.1
